Move deleteToken.ts into its own folder and unit test
