+++ server/utils/database/deleteToken/deleteToken.ts
